Switch to using `pwnlib` constants for flags.
`pwnlib` already exposes common constants in a platform-agnostic manner.
This was going to be a follow-up to the original flag formatting
support, but `pwnlib`'s functionality makes this trivial.

In some cases where `pwnlib` doesn't have a supported constant, it is
still trivial to supply a manual value (see the `MAP_SHARED_VALIDATE`
case`).